streamlit-app/requirements.txt
streamlit==1.48.1
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
